+++ docs/en/reference/types.rst
@@ -729,9 +729,7 @@ Please also notice the mapping specific footnotes for additional information.
 |                   |               +--------------------------+         |                                                          |
 |                   |               | **Oracle**               |         |                                                          |
 |                   |               +--------------------------+---------+----------------------------------------------------------+
-|                   |               | **SQL Anywhere**         | < 12    | ``DATETIME``  [15]_ [16]_                                |
-|                   |               |                          +---------+----------------------------------------------------------+
-|                   |               |                          | >= 12   | ``TIMESTAMP WITH TIME ZONE``                             |
+|                   |               | **SQL Anywhere**         | "all"   | ``TIMESTAMP WITH TIME ZONE``                             |
 +-------------------+---------------+--------------------------+---------+----------------------------------------------------------+
 | **time**          | ``\DateTime`` | **MySQL**                | *all*   | ``TIME``                                                 |
 |                   |               +--------------------------+         |                                                          |
