+++ lib/Doctrine/DBAL/Driver/AbstractSQLAnywhereDriver.php
@@ -5,14 +5,10 @@ namespace Doctrine\DBAL\Driver;
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\DBALException;
 use Doctrine\DBAL\Exception;
-use Doctrine\DBAL\Platforms\SQLAnywhere11Platform;
-use Doctrine\DBAL\Platforms\SQLAnywhere12Platform;
-use Doctrine\DBAL\Platforms\SQLAnywhere16Platform;
 use Doctrine\DBAL\Platforms\SQLAnywherePlatform;
 use Doctrine\DBAL\Schema\SQLAnywhereSchemaManager;
 use Doctrine\DBAL\VersionAwarePlatformDriver;
 use function preg_match;
-use function version_compare;
 
 /**
  * Abstract base implementation of the {@link Doctrine\DBAL\Driver} interface for SAP Sybase SQL Anywhere based drivers.
@@ -87,12 +83,6 @@ abstract class AbstractSQLAnywhereDriver implements ExceptionConverterDriver, Ve
         $version      = $majorVersion . '.' . $minorVersion . '.' . $patchVersion . '.' . $buildVersion;
 
         switch (true) {
-            case version_compare($version, '16', '>='):
-                return new SQLAnywhere16Platform();
-            case version_compare($version, '12', '>='):
-                return new SQLAnywhere12Platform();
-            case version_compare($version, '11', '>='):
-                return new SQLAnywhere11Platform();
             default:
                 return new SQLAnywherePlatform();
         }
@@ -113,7 +103,7 @@ abstract class AbstractSQLAnywhereDriver implements ExceptionConverterDriver, Ve
      */
     public function getDatabasePlatform()
     {
-        return new SQLAnywhere12Platform();
+        return new SQLAnywherePlatform();
     }
 
     /**
